move mem filter spec builder (#256)
