+++ pkg/vm/engine/disttae/logtailreplay/rows_iter.go
@@ -16,6 +16,8 @@ package logtailreplay
 
 import (
 	"bytes"
+	"fmt"
+	"github.com/matrixorigin/matrixone/pkg/objectio/ioutil"
 	"math"
 
 	"github.com/matrixorigin/matrixone/pkg/container/types"
@@ -602,14 +604,59 @@ func (p *PartitionState) NewRowsIter(ts types.TS, blockID *types.Blockid, iterDe
 	return ret
 }
 
+func buildSpec(op int, keys [][]byte) PrimaryKeyMatchSpec {
+	switch op {
+	case function.EQUAL:
+		return Exact(keys[0])
+
+	case function.PREFIX_EQ:
+		return Prefix(keys[0])
+
+	case function.IN, function.PREFIX_IN:
+		// // may be it's better to iterate rows instead.
+		// if len(f.packed) > 128 {
+		// 	return
+		// }
+		//spec = logtailreplay.InKind(f.packed, f.Op)
+		return InKind(keys, op)
+
+	case function.LESS_EQUAL, function.LESS_THAN:
+		return LessKind(keys[0], op == function.LESS_EQUAL)
+
+	case function.GREAT_EQUAL, function.GREAT_THAN:
+		return GreatKind(keys[0], op == function.GREAT_EQUAL)
+
+	case function.BETWEEN, ioutil.RangeLeftOpen,
+		ioutil.RangeRightOpen, ioutil.RangeBothOpen, function.PREFIX_BETWEEN:
+		var kind int
+		switch op {
+		case function.BETWEEN:
+			kind = 0
+		case ioutil.RangeLeftOpen:
+			kind = 1
+		case ioutil.RangeRightOpen:
+			kind = 2
+		case ioutil.RangeBothOpen:
+			kind = 3
+		case function.PREFIX_BETWEEN:
+			kind = 4
+		}
+
+		return BetweenKind(keys[0], keys[1], kind)
+	}
+
+	panic(fmt.Sprintf("build spec failed, %v, %v", op, keys))
+}
+
 func (p *PartitionState) NewPrimaryKeyIter(
 	ts types.TS,
-	spec PrimaryKeyMatchSpec,
+	op int,
+	keys [][]byte,
 ) *primaryKeyIter {
 	index := p.rowPrimaryKeyIndex
 	return &primaryKeyIter{
 		ts:           ts,
-		spec:         spec,
+		spec:         buildSpec(op, keys),
 		iter:         index.Iter(),
 		primaryIndex: index,
 		rows:         p.rows,
@@ -618,14 +665,15 @@ func (p *PartitionState) NewPrimaryKeyIter(
 
 func (p *PartitionState) NewPrimaryKeyDelIter(
 	ts *types.TS,
-	spec PrimaryKeyMatchSpec,
 	bid *types.Blockid,
+	op int,
+	keys [][]byte,
 ) *primaryKeyDelIter {
 	index := p.rowPrimaryKeyIndex
 	delIter := &primaryKeyDelIter{
 		primaryKeyIter: primaryKeyIter{
 			ts:                *ts,
-			spec:              spec,
+			spec:              buildSpec(op, keys),
 			primaryIndex:      index,
 			iter:              index.Iter(),
 			rows:              p.rows,
