+++ pkg/vm/engine/engine_util/pk_filter.go
@@ -16,6 +16,7 @@ package engine_util
 
 import (
 	"bytes"
+	"github.com/matrixorigin/matrixone/pkg/objectio/ioutil"
 
 	"github.com/matrixorigin/matrixone/pkg/common/mpool"
 	"github.com/matrixorigin/matrixone/pkg/container/types"
@@ -323,16 +324,16 @@ func ConstructBlockPKFilter(
 			unSortedSearchFunc = vector.OrderedSearchOffsetsByGreat(types.DecodeEnum(basePKFilter.LB), closed, false)
 		}
 
-	case function.BETWEEN, RangeLeftOpen, RangeRightOpen, RangeBothOpen:
+	case function.BETWEEN, ioutil.RangeLeftOpen, ioutil.RangeRightOpen, ioutil.RangeBothOpen:
 		var hint int
 		switch basePKFilter.Op {
 		case function.BETWEEN:
 			hint = 0
-		case RangeLeftOpen:
+		case ioutil.RangeLeftOpen:
 			hint = 1
-		case RangeRightOpen:
+		case ioutil.RangeRightOpen:
 			hint = 2
-		case RangeBothOpen:
+		case ioutil.RangeBothOpen:
 			hint = 3
 		}
 		switch basePKFilter.Oid {
@@ -695,7 +696,7 @@ func mergeFilters(
 				finalFilter.LB = left.LB
 				finalFilter.UB = right.LB
 				if right.Op == function.LESS_THAN {
-					finalFilter.Op = RangeRightOpen
+					finalFilter.Op = ioutil.RangeRightOpen
 				} else {
 					finalFilter.Op = function.BETWEEN
 				}
@@ -727,9 +728,9 @@ func mergeFilters(
 				finalFilter.LB = left.LB
 				finalFilter.UB = right.LB
 				if right.Op == function.LESS_THAN {
-					finalFilter.Op = RangeBothOpen
+					finalFilter.Op = ioutil.RangeBothOpen
 				} else {
-					finalFilter.Op = RangeLeftOpen
+					finalFilter.Op = ioutil.RangeLeftOpen
 				}
 				finalFilter.Valid = true
 
@@ -752,7 +753,7 @@ func mergeFilters(
 				if right.Op == function.GREAT_EQUAL {
 					finalFilter.Op = function.BETWEEN
 				} else {
-					finalFilter.Op = RangeLeftOpen
+					finalFilter.Op = ioutil.RangeLeftOpen
 				}
 				finalFilter.Valid = true
 
@@ -782,9 +783,9 @@ func mergeFilters(
 				finalFilter.LB = right.LB
 				finalFilter.UB = left.LB
 				if right.Op == function.GREAT_EQUAL {
-					finalFilter.Op = RangeRightOpen
+					finalFilter.Op = ioutil.RangeRightOpen
 				} else {
-					finalFilter.Op = RangeBothOpen
+					finalFilter.Op = ioutil.RangeBothOpen
 				}
 				finalFilter.Valid = true
